Rename a few files
In v3 folder they don't conflict with old components, so we want to call it what it is.
